Safeguard_Combined-main/backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
email-validator
python-dotenv==1.0.0
aiohttp==3.9.0
httpx==0.27.0
requests==2.31.0
websocket-client
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
polars
newspaper3k
trafilatura
readability-lxml
beautifulsoup4
playwright
feedparser
gnews
praw
telethon
google-api-python-client
emoji
tqdm
contractions
apscheduler
torch
transformers
vaderSentiment
prophet
scikit-learn
xgboost
joblib
openai==1.54.0
yfinance==0.2.32
stripe==11.1.1
